1.加入页面缓存

+++ src/main/java/com/ke/seckill/mapper/OrdersMapper.java
@@ -17,5 +17,5 @@ public interface OrdersMapper extends BaseMapper<Orders> {
      * 创建订单
      * @param orders
      */
-    long createOrder(Orders orders);
+    int createOrder(Orders orders);
 }
